+++ src/Odotocodot.OneNote.Linq/ComObjectMode.cs
@@ -1,0 +1,9 @@
+namespace Odotocodot.OneNote.Linq
+{
+    public enum ComObjectMode
+    {
+        Lazy,
+        Manual,
+        Wrap,
+    }
+}